backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
gunicorn==25.0.3
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
